review: drain the mapping repair from housekeeping, not the HID handler
CodeRabbit flagged the repair running inline in raw_hid_receive(). It is right:
a full mapping is up to 34 reports and each send_to_bridge can burn 10 retries
x the bridge timeout before giving up — on the bad link that triggers a repair
in the first place that is seconds of dead main loop inside one HID command.

Split the walk across housekeeping ticks. enable_overlays now only ARMS the
repair (arm_overlay_map_repair); overlay_map_repair_tick() sends at most
MAP_REPAIR_REPORTS_PER_TICK (2) reports per pass from a saved cursor, so each
pass costs the same as the periodic syncs while the whole mapping still lands
within a few ms of wall clock. A give-up now aborts the pass and re-latches
instead of grinding through the rest of the pool against a dead link.
Master-only (it pushes OUR tables to the slave).

Also: widen the courtyard dilation counter to int16_t. CodeRabbit called this
signed-overflow UB, which it is not — an int8_t operand promotes to int and the
conversion back is implementation-defined, as its own cited source says. The
real hazard is different and still worth closing: at radius == INT8_MAX the
int8_t counter wraps at 127 rather than exceeding `radius`, so the loop never
terminates. Unreachable today (callers pass 3, or 1 for the lang flags), and
the wider counter costs nothing.

+++ keyboards/polykybd/base/disp_array.c
@@ -733,9 +733,13 @@ void clear_line(int8_t from_x, int8_t to_x, int8_t y) {
 // reader to column-native and left the row-major overlay call site behind.
 
 // Emit the dilated clears for one detected run [run_start, run_end] on source row bmp_y.
+// `dy` is int16_t, not int8_t: at radius == INT8_MAX the int8_t counter would
+// wrap at 127 instead of exceeding `radius`, and the loop would never terminate.
+// Unreachable today (callers pass KDISP_CY_DEFAULT 3, or 1 for the lang flags),
+// but the counter costs nothing and the function takes any positive int8_t.
 static void courtyard_clear_run(int8_t x, int8_t y, int8_t bmp_y,
                                 int8_t run_start, int8_t run_end, int8_t radius) {
-    for (int8_t dy = -radius; dy <= radius; ++dy) {
+    for (int16_t dy = -(int16_t)radius; dy <= (int16_t)radius; ++dy) {
         clear_line(x + run_start - radius, x + run_end + 1 + radius, bmp_y + y + dy);
     }
 }
